+++ src/vs/workbench/contrib/localization/electron-sandbox/localization.contribution.ts
@@ -5,9 +5,7 @@
 
 import { localize } from 'vs/nls';
 import { Registry } from 'vs/platform/registry/common/platform';
-import { IWorkbenchContribution, Extensions as WorkbenchExtensions, IWorkbenchContributionsRegistry } from 'vs/workbench/common/contributions';
-import { Disposable } from 'vs/base/common/lifecycle';
-import { ExtensionsRegistry } from 'vs/workbench/services/extensions/common/extensionsRegistry';
+import { Extensions as WorkbenchExtensions, IWorkbenchContributionsRegistry } from 'vs/workbench/common/contributions';
 import { LifecyclePhase } from 'vs/workbench/services/lifecycle/common/lifecycle';
 import * as platform from 'vs/base/common/platform';
 import { IExtensionManagementService, IExtensionGalleryService, InstallOperation, ILocalExtension, InstallExtensionResult, DidUninstallExtensionEvent } from 'vs/platform/extensionManagement/common/extensionManagement';
@@ -20,18 +18,13 @@ import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { CancellationToken } from 'vs/base/common/cancellation';
 import { IPaneCompositePartService } from 'vs/workbench/services/panecomposite/browser/panecomposite';
 import { ViewContainerLocation } from 'vs/workbench/common/views';
-import { registerAction2 } from 'vs/platform/actions/common/actions';
-import { ClearDisplayLanguageAction, ConfigureDisplayLanguageAction } from 'vs/workbench/contrib/localization/browser/localizationsActions';
 import { ILocaleService } from 'vs/workbench/services/localization/common/locale';
 import { IProductService } from 'vs/platform/product/common/productService';
+import { BaseLocalizationWorkbenchContribution } from 'vs/workbench/contrib/localization/common/localization.contribution';
 
-// Register action to configure locale and related settings
-registerAction2(ConfigureDisplayLanguageAction);
-registerAction2(ClearDisplayLanguageAction);
+class NativeLocalizationWorkbenchContribution extends BaseLocalizationWorkbenchContribution {
+	private static LANGUAGEPACK_SUGGESTION_IGNORE_STORAGE_KEY = 'extensionsAssistant/languagePackSuggestionIgnore';
 
-const LANGUAGEPACK_SUGGESTION_IGNORE_STORAGE_KEY = 'extensionsAssistant/languagePackSuggestionIgnore';
-
-export class LocalizationWorkbenchContribution extends Disposable implements IWorkbenchContribution {
 	constructor(
 		@INotificationService private readonly notificationService: INotificationService,
 		@ILocaleService private readonly localeService: ILocaleService,
@@ -99,7 +92,13 @@ export class LocalizationWorkbenchContribution extends Disposable implements IWo
 	private checkAndInstall(): void {
 		const language = platform.language;
 		let locale = platform.locale ?? '';
-		const languagePackSuggestionIgnoreList = <string[]>JSON.parse(this.storageService.get(LANGUAGEPACK_SUGGESTION_IGNORE_STORAGE_KEY, StorageScope.APPLICATION, '[]'));
+		const languagePackSuggestionIgnoreList: string[] = JSON.parse(
+			this.storageService.get(
+				NativeLocalizationWorkbenchContribution.LANGUAGEPACK_SUGGESTION_IGNORE_STORAGE_KEY,
+				StorageScope.APPLICATION,
+				'[]'
+			)
+		);
 
 		if (!this.galleryService.isEnabled()) {
 			return;
@@ -203,7 +202,7 @@ export class LocalizationWorkbenchContribution extends Disposable implements IWo
 								run: () => {
 									languagePackSuggestionIgnoreList.push(fullLocale);
 									this.storageService.store(
-										LANGUAGEPACK_SUGGESTION_IGNORE_STORAGE_KEY,
+										NativeLocalizationWorkbenchContribution.LANGUAGEPACK_SUGGESTION_IGNORE_STORAGE_KEY,
 										JSON.stringify(languagePackSuggestionIgnoreList),
 										StorageScope.APPLICATION,
 										StorageTarget.USER
@@ -229,55 +228,4 @@ export class LocalizationWorkbenchContribution extends Disposable implements IWo
 }
 
 const workbenchRegistry = Registry.as<IWorkbenchContributionsRegistry>(WorkbenchExtensions.Workbench);
-workbenchRegistry.registerWorkbenchContribution(LocalizationWorkbenchContribution, LifecyclePhase.Eventually);
-
-ExtensionsRegistry.registerExtensionPoint({
-	extensionPoint: 'localizations',
-	defaultExtensionKind: ['ui', 'workspace'],
-	jsonSchema: {
-		description: localize('vscode.extension.contributes.localizations', "Contributes localizations to the editor"),
-		type: 'array',
-		default: [],
-		items: {
-			type: 'object',
-			required: ['languageId', 'translations'],
-			defaultSnippets: [{ body: { languageId: '', languageName: '', localizedLanguageName: '', translations: [{ id: 'vscode', path: '' }] } }],
-			properties: {
-				languageId: {
-					description: localize('vscode.extension.contributes.localizations.languageId', 'Id of the language into which the display strings are translated.'),
-					type: 'string'
-				},
-				languageName: {
-					description: localize('vscode.extension.contributes.localizations.languageName', 'Name of the language in English.'),
-					type: 'string'
-				},
-				localizedLanguageName: {
-					description: localize('vscode.extension.contributes.localizations.languageNameLocalized', 'Name of the language in contributed language.'),
-					type: 'string'
-				},
-				translations: {
-					description: localize('vscode.extension.contributes.localizations.translations', 'List of translations associated to the language.'),
-					type: 'array',
-					default: [{ id: 'vscode', path: '' }],
-					items: {
-						type: 'object',
-						required: ['id', 'path'],
-						properties: {
-							id: {
-								type: 'string',
-								description: localize('vscode.extension.contributes.localizations.translations.id', "Id of VS Code or Extension for which this translation is contributed to. Id of VS Code is always `vscode` and of extension should be in format `publisherId.extensionName`."),
-								pattern: '^((vscode)|([a-z0-9A-Z][a-z0-9A-Z-]*)\\.([a-z0-9A-Z][a-z0-9A-Z-]*))$',
-								patternErrorMessage: localize('vscode.extension.contributes.localizations.translations.id.pattern', "Id should be `vscode` or in format `publisherId.extensionName` for translating VS code or an extension respectively.")
-							},
-							path: {
-								type: 'string',
-								description: localize('vscode.extension.contributes.localizations.translations.path', "A relative path to a file containing translations for the language.")
-							}
-						},
-						defaultSnippets: [{ body: { id: '', path: '' } }],
-					},
-				}
-			}
-		}
-	}
-});
+workbenchRegistry.registerWorkbenchContribution(NativeLocalizationWorkbenchContribution, LifecyclePhase.Eventually);
